Got first externalized Tooltip & semantic expansions generation working. Use AbstractVDServerTask to handle common preparation and threading/cancellation infrastructure.
